+++ py2max/tests/test_table.py
@@ -1,0 +1,12 @@
+
+from .. import Patcher
+
+
+def test_table():
+    p = Patcher('outputs/test_table.maxpat')
+    p.add_table('box', array=list(range(128)))
+    p.save()
+
+
+if __name__ == '__main__':
+    test_table()