+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## 0.1
 
-- added coll object and tests.
+- added `from_file` classmethod to Patcher to populate object from .maxpat file.
+- added coll, dict and table objects and tests
 
 - add some tests which try to use generic layout algorithm in Networkx but the results are quite terrible using the so probably better to just create something fit-for-purpose.
 - added gen subpatcher